defect-classifier/requirements.txt
streamlit
numpy
pillow
torch
torchvision